% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TEAM_IDS}
\alias{TEAM_IDS}
\title{Basketball Reference IDs for Active NBA teams.}
\format{
A data frame with 30 rows and 2 variables:
\describe{
\item{Franchise}{franchise name}
\item{TeamId}{ID of team used in Basketball Reference URL}
}
}
\source{
Hyperlinks for each team at \url{https://www.basketball-reference.com/teams/}
}
\usage{
TEAM_IDS
}
\description{
A dataset containing the franchise names and IDs used by Basketball Reference to
uniquely identify each active team.
}
\keyword{datasets}
