% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams.franchise_index}
\alias{teams.franchise_index}
\title{Index of all seasons for given NBA franchises.}
\usage{
teams.franchise_index(team_ids = NULL)
}
\arguments{
\item{team_ids}{A vector of Basketball Reference team IDs (see \code{\link{TEAM_IDS}})}
}
\description{
Returns the row binded "Seasons" tables as shown on individual team pages
e.g. \url{https://www.basketball-reference.com/teams/ATL/}.
}
\details{
\code{teams.franchise_index} adds the convenience variable MadePlayoffs,
representing whether the team made the playoffs in the corresponding season.
This is determined by parsing the "*" symbol off the end of the Team variable
such that there are no asterisks in the resulting Team name.
}
