% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams.team_index}
\alias{teams.team_index}
\title{Index of all active NBA franchises.}
\usage{
teams.team_index(include_total_history = FALSE)
}
\arguments{
\item{include_total_history}{If TRUE, return data frame with all historical
counterparts of each franchise listed as separate rows.}
}
\description{
Returns the "Active Franchises" table as shown at
\url{https://www.basketball-reference.com/teams/}.
}
\details{
A TeamId variable is added that represents the Basketball Reference
unique identifier for that team as shown in the team URL e.g. Atlanta - ATL
(\url{https://www.basketball-reference.com/teams/ATL/}).
}
